refactor: browse page components
