+++ apps/website/src/app/browse/loading.tsx
@@ -1,10 +1,7 @@
 import { SeriesList } from "@easyshell/problems/data/series"
 
-import { Progress } from "@/components/ui/progress"
-
-import { ProblemListSkeleton } from "./client"
-
-import Link from "next/link"
+import { ProblemListSkeleton } from "./_components/problems"
+import { SeriesCardSkeleton } from "./_components/series"
 
 export default function Page() {
   return (
@@ -26,33 +23,5 @@ export default function Page() {
       </div>
       <ProblemListSkeleton />
     </div>
-  )
-}
-
-function SeriesCardSkeleton({
-  series,
-}: {
-  series: (typeof SeriesList)[number]
-}) {
-  return (
-    <Link
-      href={`/series/${series.slug}`}
-      key={series.slug}
-      className="flex w-60 flex-col overflow-hidden rounded-xl border transition-colors hover:bg-neutral-50 dark:bg-neutral-950"
-    >
-      <div className="h-18 animate-pulse bg-neutral-100 dark:bg-stone-900"></div>
-      <div className="flex flex-col px-4 py-2">
-        <div className="flex items-center justify-between">
-          <div className="font-clash-display font-semibold">{series.name}</div>
-        </div>
-        <div className="text-justify text-sm text-neutral-600">
-          {series.description}
-        </div>
-        <Progress
-          className="my-2 animate-pulse bg-emerald-100 dark:bg-emerald-950 dark:*:bg-emerald-800"
-          value={0}
-        />
-      </div>
-    </Link>
   )
 }
